requirements.txt
django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
python-dotenv==1.0.1
gunicorn==22.0.0
psycopg2-binary==2.9.9 